# fastingmate-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Dokumentasi API untuk autentikasi pengguna",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/user-targets": {
            "get": {
                "tags": [
                    "User Targets"
                ],
                "summary": "Menampilkan semua target pengguna",
                "operationId": "908befe8aee8cccdac55d1c25da82262",
                "responses": {
                    "200": {
                        "description": "Daftar target pengguna berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "target_weight": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 65
                                            },
                                            "target_daily_calories": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 2000
                                            },
                                            "target_protein": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 100
                                            },
                                            "target_carbs": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 250
                                            },
                                            "target_fat": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 50
                                            },
                                            "target_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-06-30"
                                            },
                                            "is_achieved": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Targets"
                ],
                "summary": "Menambahkan target baru untuk pengguna",
                "operationId": "a6b246633281478cfc45c4507bc4d92f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "target_weight",
                                    "target_daily_calories",
                                    "target_protein",
                                    "target_carbs",
                                    "target_fat",
                                    "target_date"
                                ],
                                "properties": {
                                    "target_weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 65
                                    },
                                    "target_daily_calories": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000
                                    },
                                    "target_protein": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "target_carbs": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    },
                                    "target_fat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "target_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Target berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-targets/{id}": {
            "delete": {
                "tags": [
                    "User Targets"
                ],
                "summary": "Menghapus target berdasarkan ID",
                "operationId": "2a3d7c91717f2d4905c2bcee809103b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target berhasil dihapus"
                    },
                    "404": {
                        "description": "Target tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/weight-logs": {
            "get": {
                "tags": [
                    "Weight Logs"
                ],
                "summary": "Menampilkan daftar catatan berat badan pengguna",
                "operationId": "2330a2a1cb23896d0e57194e6f9194b2",
                "responses": {
                    "200": {
                        "description": "Daftar catatan berat badan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "weight": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 70.5
                                            },
                                            "recorded_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-03-21"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2025-03-21 10:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Weight Logs"
                ],
                "summary": "Menambahkan catatan berat badan baru",
                "operationId": "96b3f10105dad08971c59ad876483f78",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "weight",
                                    "recorded_date"
                                ],
                                "properties": {
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 70.5
                                    },
                                    "recorded_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-21"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catatan berat badan berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/weight-logs/{id}": {
            "delete": {
                "tags": [
                    "Weight Logs"
                ],
                "summary": "Menghapus catatan berat badan berdasarkan ID",
                "operationId": "80a33004ae0478460fb51d4343cf30b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catatan berat badan berhasil dihapus"
                    },
                    "404": {
                        "description": "Catatan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <your-token>)"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}